AddisNet/requirements.txt
# AddisNet - Smart City Platform for Addis Ababa
# Django, Firebase, Real-time Analytics, Community Engagement

Django>=5.0
djangorestframework>=3.14
firebase-admin>=6.0
Pillow>=10.0
django-cors-headers>=4.0
channels>=4.0
channels-redis>=4.0
psycopg2-binary>=2.9  # Optional: for PostgreSQL
gunicorn>=21.0
python-decouple>=3.8
requests>=2.31
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3  # For predictive analytics stub
